requirements.txt
numpy
nltk
transformers
torch
python-dotenv
pinecone